+++ stroom-core-client-widget/src/main/java/stroom/editor/client/view/EditorViewImpl.java
@@ -20,7 +20,6 @@ import stroom.editor.client.event.FormatEvent;
 import stroom.editor.client.event.FormatEvent.FormatHandler;
 import stroom.editor.client.model.XmlFormatter;
 import stroom.editor.client.presenter.Action;
-import stroom.editor.client.presenter.EditorUiHandlers;
 import stroom.editor.client.presenter.EditorView;
 import stroom.editor.client.presenter.Option;
 import stroom.util.shared.DefaultLocation;
@@ -31,11 +30,8 @@ import stroom.util.shared.TextRange;
 import stroom.widget.contextmenu.client.event.ContextMenuEvent;
 import stroom.widget.tab.client.view.ResizeObserver;
 
-import com.google.gwt.core.client.GWT;
 import com.google.gwt.core.client.Scheduler;
 import com.google.gwt.dom.client.NativeEvent;
-import com.google.gwt.dom.client.Style.Unit;
-import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.dom.client.KeyDownEvent;
 import com.google.gwt.event.dom.client.KeyDownHandler;
 import com.google.gwt.event.dom.client.MouseDownEvent;
@@ -45,10 +41,9 @@ import com.google.gwt.event.shared.GwtEvent;
 import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.uibinder.client.UiField;
-import com.google.gwt.user.client.ui.Button;
 import com.google.gwt.user.client.ui.FlowPanel;
 import com.google.gwt.user.client.ui.Widget;
-import com.gwtplatform.mvp.client.ViewWithUiHandlers;
+import com.gwtplatform.mvp.client.ViewImpl;
 import edu.ycp.cs.dh.acegwt.client.ace.AceAnnotationType;
 import edu.ycp.cs.dh.acegwt.client.ace.AceEditorCursorPosition;
 import edu.ycp.cs.dh.acegwt.client.ace.AceEditorMode;
@@ -69,11 +64,10 @@ import javax.inject.Inject;
  * functionality such as formatting, styling, line numbers and warning/error
  * markers.
  */
-public class EditorViewImpl extends ViewWithUiHandlers<EditorUiHandlers> implements EditorView {
+public class EditorViewImpl extends ViewImpl implements EditorView {
 
     private static final IndicatorPopup indicatorPopup = new IndicatorPopup();
     private static final boolean SHOW_INDICATORS_DEFAULT = false;
-    private static volatile Binder binder;
     private final Action formatAction;
     private final Option stylesOption;
     private final Option lineNumbersOption;
@@ -92,12 +86,6 @@ public class EditorViewImpl extends ViewWithUiHandlers<EditorUiHandlers> impleme
     Editor editor;
     @UiField
     RightBar rightBar;
-    @UiField
-    FlowPanel filterButtons;
-    @UiField
-    Button filterInactive;
-    @UiField
-    Button filterActive;
 
     private IndicatorLines indicators;
     private AceEditorMode mode = AceEditorMode.XML;
@@ -108,24 +96,6 @@ public class EditorViewImpl extends ViewWithUiHandlers<EditorUiHandlers> impleme
     public EditorViewImpl(final Binder binder) {
         layout = binder.createAndBindUi(this);
         ResizeObserver.observe(layout.getElement(), element -> doLayout(SHOW_INDICATORS_DEFAULT));
-
-        filterButtons.addDomHandler(event -> {
-            if ((event.getNativeButton() & NativeEvent.BUTTON_LEFT) != 0) {
-                if (getUiHandlers() != null) {
-                    getUiHandlers().changeFilterSettings();
-                }
-            }
-        }, ClickEvent.getType());
-
-        filterButtons.setWidth(ScrollbarMetrics.getVerticalScrollBarWidth() + "px");
-        filterButtons.setHeight(ScrollbarMetrics.getHorizontalScrollBarWidth() + "px");
-
-        final int left = ((ScrollbarMetrics.getVerticalScrollBarWidth() - 10) / 2);
-        final int top = ((ScrollbarMetrics.getHorizontalScrollBarWidth() - 10) / 2);
-        filterInactive.getElement().getStyle().setLeft(left, Unit.PX);
-        filterActive.getElement().getStyle().setLeft(left, Unit.PX);
-        filterInactive.getElement().getStyle().setTop(top, Unit.PX);
-        filterActive.getElement().getStyle().setTop(top, Unit.PX);
 
         formatAction = new Action("Format", false, this::format);
 
@@ -152,7 +122,7 @@ public class EditorViewImpl extends ViewWithUiHandlers<EditorUiHandlers> impleme
                 "Highlight Active Line", true, true, (on) -> editor.setHighlightActiveLine(on));
 
         editor.getElement().setClassName("editor");
-        editor.addDomHandler(event -> handleMouseDown(event), MouseDownEvent.getType());
+        editor.addDomHandler(this::handleMouseDown, MouseDownEvent.getType());
         rightBar.setEditor(editor);
     }
 
@@ -170,17 +140,12 @@ public class EditorViewImpl extends ViewWithUiHandlers<EditorUiHandlers> impleme
         return layout;
     }
 
-    private void updateGutter(final boolean showLineNumbers) {
-        editor.setShowGutter(showLineNumbers);
-    }
-
     private void doLayout() {
         doLayout(indicatorsOption.isOn());
     }
 
     private void doLayout(final boolean showIndicators) {
         rightBar.render(indicators, showIndicators);
-//        layout.setWidgetSize(rightBar, rightBar.getWidth());
         editor.onResize();
     }
 
@@ -261,7 +226,7 @@ public class EditorViewImpl extends ViewWithUiHandlers<EditorUiHandlers> impleme
                         }
 
                         final Severity severity = error.getSeverity();
-                        AceAnnotationType annotationType = AceAnnotationType.INFO;
+                        AceAnnotationType annotationType;
                         switch (severity) {
                             case INFO:
                                 annotationType = AceAnnotationType.INFO;
@@ -467,17 +432,6 @@ public class EditorViewImpl extends ViewWithUiHandlers<EditorUiHandlers> impleme
     @Override
     public Option getHighlightActiveLineOption() {
         return highlightActiveLineOption;
-    }
-
-    @Override
-    public void showFilterButton(final boolean show) {
-        filterActive.setVisible(false);
-        filterButtons.setVisible(show);
-    }
-
-    @Override
-    public void setFilterActive(final boolean active) {
-        filterActive.setVisible(active);
     }
 
     @Override